[Chore] Fix lint errors caused by casting int to int32 (#2368)

+++ ray-operator/controllers/ray/common/pod.go
@@ -264,14 +264,14 @@ func initLivenessAndReadinessProbe(rayContainer *corev1.Container, rayNodeType r
 	}
 
 	if rayContainer.LivenessProbe == nil {
-		probeTimeout := utils.DefaultLivenessProbeTimeoutSeconds
+		probeTimeout := int32(utils.DefaultLivenessProbeTimeoutSeconds)
 		if rayNodeType == rayv1.HeadNode {
-			probeTimeout = utils.DefaultHeadLivenessProbeTimeoutSeconds
+			probeTimeout = int32(utils.DefaultHeadLivenessProbeTimeoutSeconds)
 		}
 
 		rayContainer.LivenessProbe = &corev1.Probe{
 			InitialDelaySeconds: utils.DefaultLivenessProbeInitialDelaySeconds,
-			TimeoutSeconds:      int32(probeTimeout),
+			TimeoutSeconds:      probeTimeout,
 			PeriodSeconds:       utils.DefaultLivenessProbePeriodSeconds,
 			SuccessThreshold:    utils.DefaultLivenessProbeSuccessThreshold,
 			FailureThreshold:    utils.DefaultLivenessProbeFailureThreshold,
@@ -280,13 +280,13 @@ func initLivenessAndReadinessProbe(rayContainer *corev1.Container, rayNodeType r
 	}
 
 	if rayContainer.ReadinessProbe == nil {
-		probeTimeout := utils.DefaultReadinessProbeTimeoutSeconds
+		probeTimeout := int32(utils.DefaultReadinessProbeTimeoutSeconds)
 		if rayNodeType == rayv1.HeadNode {
-			probeTimeout = utils.DefaultHeadReadinessProbeTimeoutSeconds
+			probeTimeout = int32(utils.DefaultHeadReadinessProbeTimeoutSeconds)
 		}
 		rayContainer.ReadinessProbe = &corev1.Probe{
 			InitialDelaySeconds: utils.DefaultReadinessProbeInitialDelaySeconds,
-			TimeoutSeconds:      int32(probeTimeout),
+			TimeoutSeconds:      probeTimeout,
 			PeriodSeconds:       utils.DefaultReadinessProbePeriodSeconds,
 			SuccessThreshold:    utils.DefaultReadinessProbeSuccessThreshold,
 			FailureThreshold:    utils.DefaultReadinessProbeFailureThreshold,
